add integration tests for browser extension discoverability changes
