+++ client/web/src/repo/actions/GoToCodeHostAction.tsx
@@ -3,8 +3,7 @@ import { upperFirst } from 'lodash'
 import BitbucketIcon from 'mdi-react/BitbucketIcon'
 import ExportIcon from 'mdi-react/ExportIcon'
 import GithubIcon from 'mdi-react/GithubIcon'
-import PlusThickIcon from 'mdi-react/PlusThickIcon'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useCallback, useEffect, useState } from 'react'
 import { merge, Observable, of } from 'rxjs'
 import { catchError, distinctUntilChanged, startWith, switchMap } from 'rxjs/operators'
 import { PhabricatorIcon } from '../../../../shared/src/components/icons' // TODO: Switch mdi icon
@@ -16,9 +15,8 @@ import { RevisionSpec, FileSpec } from '../../../../shared/src/util/url'
 import { ExternalLinkFields } from '../../graphql-operations'
 import { useEventObservable, useObservable } from '../../../../shared/src/util/useObservable'
 import GitlabIcon from 'mdi-react/GitlabIcon'
-import { SourcegraphIcon } from '../../auth/icons'
 import { eventLogger } from '../../tracking/eventLogger'
-import { PopoverContainer } from '../../components/PopoverContainer'
+import { InstallExtensionPopover } from './InstallExtensionPopover'
 
 interface GoToCodeHostPopoverProps {
     canShowPopover: boolean
@@ -201,7 +199,7 @@ export const GoToCodeHostAction: React.FunctionComponent<Props> = props => {
             </ButtonLink>
 
             {showPopover && (
-                <CodeHostExtensionPopover
+                <InstallExtensionPopover
                     url={url}
                     serviceType={externalURL.serviceType}
                     onClose={onClose}
@@ -214,86 +212,7 @@ export const GoToCodeHostAction: React.FunctionComponent<Props> = props => {
     )
 }
 
-interface CodeHostExtensionPopoverProps {
-    url: string
-    serviceType: string | null
-    onClose: () => void
-    onRejection: () => void
-    onClickInstall: () => void
-    targetID: string
-}
-
-export const CodeHostExtensionPopover: React.FunctionComponent<CodeHostExtensionPopoverProps> = ({
-    url,
-    serviceType,
-    onClose,
-    onRejection,
-    onClickInstall,
-    targetID,
-}) => {
-    const { displayName, icon } = serviceTypeDisplayNameAndIcon(serviceType)
-    const Icon = icon || ExportIcon
-
-    return (
-        <PopoverContainer
-            onClose={onClose}
-            targetID={targetID}
-            popperOptions={useMemo(
-                () => ({
-                    placement: 'bottom-start' as const,
-                    modifiers: [
-                        {
-                            name: 'offset',
-                            options: {
-                                offset: [64, 4],
-                            },
-                        },
-                    ],
-                }),
-                []
-            )}
-        >
-            {modalBodyReference => (
-                <div
-                    ref={modalBodyReference as React.MutableRefObject<HTMLDivElement>}
-                    className="extension-permission-modal p-4 web-content text-wrap border shadow"
-                >
-                    <h3 className="mb-0">Take Sourcegraph's code intelligence to {displayName}!</h3>
-                    <p className="py-3">
-                        Install Sourcegraph browser extension to get code intelligence while browsing files and reading
-                        PRs on {displayName}.
-                    </p>
-
-                    <div className="mx-auto code-host-action__graphic-container d-flex justify-content-between align-items-center">
-                        <SourcegraphIcon size={48} />
-                        <PlusThickIcon size={20} className="code-host-action__plus-icon" />
-                        <Icon size={56} />
-                    </div>
-
-                    <div className="d-flex justify-content-end">
-                        <ButtonLink className="btn btn-outline-secondary mr-2" onSelect={onRejection} to={url}>
-                            No, thanks
-                        </ButtonLink>
-
-                        <ButtonLink className="btn btn-outline-secondary mr-2" onSelect={onClose} to={url}>
-                            Remind me later
-                        </ButtonLink>
-
-                        <ButtonLink
-                            className="btn btn-primary mr-2"
-                            onSelect={onClickInstall}
-                            to="/help/integration/browser_extension"
-                        >
-                            Install browser extension
-                        </ButtonLink>
-                    </div>
-                </div>
-            )}
-        </PopoverContainer>
-    )
-}
-
-function serviceTypeDisplayNameAndIcon(
+export function serviceTypeDisplayNameAndIcon(
     serviceType: string | null
 ): { displayName: string; icon?: React.ComponentType<{ className?: string }> } {
     switch (serviceType) {
