+++ client/web/src/repo/actions/InstallExtensionPopover.story.tsx
